Change to send hot loaded code directly from server
Previously, hot loaded code is sent through internal web socket
client. However, it is revealed that it makes difficult to split hot
loader as a standalone library.

For such design, a user of the library need to initialize not only
server as a middleware of lack but also the internal client. This is
because the client needs to know a port number of server while there
is no way to inform it to the client when initializing the
middleware.